agents/requirements.txt
# Requirements for the agentic misinformation system (without Google Cloud)
python-dotenv
requests
